requirements.txt
notion-client
alpaca-py
